Final_project/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
Markdown==3.2.1
django-filter==2.2.0
django_crispy_forms==1.9.0
#for myscripts
openpyxl==3.0.3
requests==2.23.0
beautifulsoup4==4.8.2